requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
paho-mqtt==2.0.0
asyncio-mqtt==0.16.1 